requirements.txt
apache-airflow
nltk
pandas
scikit-learn
requests
beautifulsoup4
